requirements.txt
openpyxl==3.1.2
pandas==2.2.1
pymysql==1.1.0
python-dotenv==1.0.1